functions: optimization: get the DOI for a file -> avoid unecessary web requests. Adding function unLike to unlike a post.
hosts: adding function getDoi, to get the DOI for a file. Also, now the
parsing of the data returns the origin url of a post.

know.py: adding preferences: can restore window state.
Adding images for the actions.
Adding action to calculate percentage match.
Adding actions to like and unlike a post.
Adding action to open a post in browser.