advanced_search: the search now almost works. Normal queries work, * not implemented yet.
gui.py: very important improvements to deal with the
models/views/proxys. Now, each tab has its own model/view/proxy. All the
tabs are generated in the same way, and they are given a database query
to display.
The like action is now done in gui.py, not in functions.py anymore.

hosts.py: topic_simple is now modified with 2 blank spaces: one at the
beginning, and one at the end, to make the LIKE queries easier in the
searches.

view.py: to define the view of each new tab