HUGE commit.
Basically, the async version of the worker is ready. It means parsing
the articles is done with a small number of threads, each thread using
several requestsFuture to download asynchronously the articles.

Also, the gui has now a function updateView, which update the view
without loosing the selection. The user can so browse the articles while
refreshing the articles.

test_* files have beed added to test the program. These files test the
worker and the getData from hosts.py.