gui.py: simplification of the code of checkThreads. I added a busy progress bar for the calculation of the percentages.
hosts.py: now, the company is given to getData and getDoi, to avoid multiple class
to getJournals. getJournals opens files, and it was causing a bug: too
many files opened.

worker.py: modification to get here the company, as a string, to give it to hosts.getData
and hosts.getDoi. This way, hosts.getJournals is called only in the worker. It
greatly reduces the number of calls to hosts.getJournals.

Other files: minor modifications, mainly to adapt to the change related
to the company.