hosts.py: creation of the updateData function. Called by a Worker object, it deals w/ how the program should handle an "incomplete" entry. It basically avoids unecessary requests to the servers, and speeds up the refresh of the new articles.
Also, a boolean has been added to the journals conf files, to let the
progam know if it has to care about the graphical abstract. If this
boolean is False, ChemBrows will not try to dl the pictures for this
journal.

worker.py: fixed a bug for the Thieme company, which uses a HTTPS page
for its RSS pages. The fix disables the SSL verifications for this
company. Also, worker.py has been updated to use updateData, and
makes a lot less requests to the servers. The speed of the refresh
is largely improved.