Addition of the license and copyright files. ChemBrows now has a license and is almost ready for diffusion.
The text of the tutorial is ready. The tutorial module has been modified
to handle only one file tuto.txt. All the slides are in this file.

Some modifications have been made on the style. A white border appears
when a button is clicked.

HUGE improvements and simplification of the module line_clear. The code
is shorter, and fixes a bug: the text could go behind the button, it was
ugly.