functions.py: merged unidecodePerso & simpleChar. Simplification of the building of the SQL query, for the authors. Now uses the author_simple field. Also fix a bug related to special characters in the topic fields.
worker: fix for a huge bug: author_simple was not inserted into the
database.

gui.py: modification of cleanDb to fix the author_simple bug, for the
users. Users w/ a db w/ the author_simple field can simply run this
function to update their db. If the db had empty author_simple fields,
this function will also correct it.